# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Interview Practice Platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: Auth
    description: Browser and mobile authentication endpoints.
  - name: gemini
    description: Gemini AI chat and image operations
  - name: interviews
    description: Interview session management
  - name: analytics
    description: User performance analytics and dashboard
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect on success or failure.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'

  /gemini/conversations:
    get:
      operationId: listGeminiConversations
      tags: [gemini]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeminiConversation"
    post:
      operationId: createGeminiConversation
      tags: [gemini]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeminiConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiConversation"
  /gemini/conversations/{id}:
    get:
      operationId: getGeminiConversation
      tags: [gemini]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiError"
    delete:
      operationId: deleteGeminiConversation
      tags: [gemini]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiError"
  /gemini/conversations/{id}/messages:
    get:
      operationId: listGeminiMessages
      tags: [gemini]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeminiMessage"
    post:
      operationId: sendGeminiMessage
      tags: [gemini]
      summary: Send a message and receive an AI response (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeminiMessageInput"
      responses:
        "200":
          description: SSE stream of assistant response chunks
          content:
            text/event-stream: {}
  /gemini/generate-image:
    post:
      operationId: generateGeminiImage
      tags: [gemini]
      summary: Generate an image from a text prompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeminiImageInput"
      responses:
        "200":
          description: Generated image data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiImageOutput"

  /interviews:
    get:
      operationId: listInterviews
      tags: [interviews]
      summary: List user's interview sessions
      responses:
        "200":
          description: List of interviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Interview"
        "401":
          description: Unauthorized
    post:
      operationId: createInterview
      tags: [interviews]
      summary: Start a new interview session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewInput"
      responses:
        "201":
          description: Created interview session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"
        "401":
          description: Unauthorized
  /interviews/{id}:
    get:
      operationId: getInterview
      tags: [interviews]
      summary: Get interview session with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Interview with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewWithMessages"
        "401":
          description: Unauthorized
        "404":
          description: Not found
  /interviews/{id}/end:
    post:
      operationId: endInterview
      tags: [interviews]
      summary: End an interview session and generate analytics
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ended interview with analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewEndResult"
        "401":
          description: Unauthorized
        "404":
          description: Not found
  /interviews/{id}/message:
    post:
      operationId: sendInterviewMessage
      tags: [interviews]
      summary: Send a message to the AI interviewer (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewMessageInput"
      responses:
        "200":
          description: SSE stream of AI interviewer response
          content:
            text/event-stream: {}
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /analytics:
    get:
      operationId: getUserAnalytics
      tags: [analytics]
      summary: Get user performance analytics
      responses:
        "200":
          description: User analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserAnalytics"
        "401":
          description: Unauthorized
  /dashboard:
    get:
      operationId: getDashboard
      tags: [analytics]
      summary: Get dashboard summary
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
        "401":
          description: Unauthorized

components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    GeminiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    GeminiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    GeminiConversationInput:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    GeminiMessageInput:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    GeminiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/GeminiMessage"
      required:
        - id
        - title
        - createdAt
        - messages
    GeminiImageInput:
      type: object
      properties:
        prompt:
          type: string
      required:
        - prompt
    GeminiImageOutput:
      type: object
      properties:
        b64_json:
          type: string
        mimeType:
          type: string
      required:
        - b64_json
        - mimeType
    GeminiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Interview:
      type: object
      required: [id, userId, role, difficulty, interviewType, status, conversationId, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        role:
          type: string
        difficulty:
          type: string
        interviewType:
          type: string
        status:
          type: string
        conversationId:
          type: integer
        confidenceScore:
          type: ["number", "null"]
        technicalScore:
          type: ["number", "null"]
        communicationScore:
          type: ["number", "null"]
        weakTopics:
          type: ["string", "null"]
        improvementSuggestions:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        endedAt:
          type: ["string", "null"]
          format: date-time
    InterviewInput:
      type: object
      required: [role, difficulty, interviewType]
      properties:
        role:
          type: string
        difficulty:
          type: string
        interviewType:
          type: string
    InterviewMessage:
      type: object
      required: [id, interviewId, role, content, createdAt]
      properties:
        id:
          type: integer
        interviewId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    InterviewWithMessages:
      type: object
      required: [id, userId, role, difficulty, interviewType, status, conversationId, createdAt, messages]
      properties:
        id:
          type: integer
        userId:
          type: string
        role:
          type: string
        difficulty:
          type: string
        interviewType:
          type: string
        status:
          type: string
        conversationId:
          type: integer
        confidenceScore:
          type: ["number", "null"]
        technicalScore:
          type: ["number", "null"]
        communicationScore:
          type: ["number", "null"]
        weakTopics:
          type: ["string", "null"]
        improvementSuggestions:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        endedAt:
          type: ["string", "null"]
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/InterviewMessage"
    InterviewMessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
    InterviewEndResult:
      type: object
      required: [interview]
      properties:
        interview:
          $ref: "#/components/schemas/Interview"

    UserAnalytics:
      type: object
      required: [totalInterviews, avgConfidenceScore, avgTechnicalScore, avgCommunicationScore, topWeakTopics, recentSessions]
      properties:
        totalInterviews:
          type: integer
        completedInterviews:
          type: integer
        avgConfidenceScore:
          type: ["number", "null"]
        avgTechnicalScore:
          type: ["number", "null"]
        avgCommunicationScore:
          type: ["number", "null"]
        topWeakTopics:
          type: array
          items:
            type: string
        recentSessions:
          type: array
          items:
            $ref: "#/components/schemas/Interview"

    Dashboard:
      type: object
      required: [totalInterviews, completedInterviews, avgScore, recentInterviews, roleBreakdown, difficultyBreakdown]
      properties:
        totalInterviews:
          type: integer
        completedInterviews:
          type: integer
        avgScore:
          type: ["number", "null"]
        recentInterviews:
          type: array
          items:
            $ref: "#/components/schemas/Interview"
        roleBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/RoleCount"
        difficultyBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/DifficultyCount"
    RoleCount:
      type: object
      required: [role, count]
      properties:
        role:
          type: string
        count:
          type: integer
    DifficultyCount:
      type: object
      required: [difficulty, count]
      properties:
        difficulty:
          type: string
        count:
          type: integer