# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Interview Practice Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { DifficultyCount } from "./difficultyCount";
import type { Interview } from "./interview";
import type { RoleCount } from "./roleCount";

export interface Dashboard {
  totalInterviews: number;
  completedInterviews: number;
  /** @nullable */
  avgScore: number | null;
  recentInterviews: Interview[];
  roleBreakdown: RoleCount[];
  difficultyBreakdown: DifficultyCount[];
}